Update Group 9 project.py
Found a headache naming issue, that R0 is used as nr of individuals removed at day 0 as well as R0 value. Now the R0 is nr of individuals removed at day 0 and R_0_virus is the R0 value of the virus. Hopefully I found all the places where this miss of naming has happened. 

Also added introduction pieces for the user to make the user experience better and more fluid

tweeked main function structure and and made it more clear. If you want to move a variable from one func to another use return var_a at the end of func_a and call it in the main function as var_a = func_a() and then you can call it in func_b as func_b(var_a)